requirements.txt
fastapi==0.95.1
nltk==3.8.1
python-dotenv==1.0.0
uvicorn==0.22.0
